+++ src/main/java/com/partymusicapp/service/UserService.java
@@ -29,5 +29,5 @@ public interface UserService {
 
     List<UserDTO> getAllUsers();
 
-    int getUserCount();
+    long getUserCount();
 }
